Move try_path to toil.lib.io
